Add Mull to the privileged FIDO2 list
Mull is a fork of Fennec, with privacy and security patches from Arkenfox, and a few of their own.

First key is the DivestOS build, the official build.
Second key is the F-Droid build, signed by F-Droid. Both are otherwise the same and are both official forms of distribution of the browser.